+++ services/resources-getter.js
@@ -15,6 +15,11 @@ function ResourcesGetter(model, opts, params) {
   var segmentScope;
   var segmentWhere;
 
+  var fields = _.clone(schema.fields);
+  var associations = _.clone(model.associations);
+  var hasSearchFields = schema.searchFields && _.isArray(schema.searchFields);
+  var searchAssociationFields;
+
   var fieldNamesRequested = (function() {
     if (!params.fields || !params.fields[model.name]) { return null; }
 
@@ -39,11 +44,46 @@ function ResourcesGetter(model, opts, params) {
       associationsForQuery);
   })();
 
+  function selectSearchFields() {
+    var searchFields = _.clone(schema.searchFields);
+    searchAssociationFields = _.remove(searchFields, function (field) {
+      return field.indexOf('.') !== -1;
+    });
+
+    _.remove(fields, function (field) {
+      return !_.includes(schema.searchFields, field.field);
+    });
+
+    var searchAssociationNames = _.map(searchAssociationFields,
+      function (association) { return association.split('.')[0]; });
+    associations = _.pick(associations, searchAssociationNames);
+
+    // NOTICE: Compute warnings to help developers to configure the
+    //         searchFields.
+    var fieldsSimpleNotFound = _.xor(searchFields,
+      _.map(fields, function (field) { return field.field; }));
+    var fieldsAssociationNotFound = _.xor(
+      _.map(searchAssociationFields, function (association) {
+        return association.split('.')[0];
+      }), _.keys(associations));
+
+    if (fieldsSimpleNotFound.length) {
+      Interface.logger.warn('Cannot find the fields [' + fieldsSimpleNotFound +
+        '] while searching records in model ' + model.name + '.');
+    }
+
+    if (fieldsAssociationNotFound.length) {
+      Interface.logger.warn('Cannot find the associations [' +
+        fieldsAssociationNotFound + '] while searching records in model ' +
+        model.name + '.');
+    }
+  }
+
   function handleSearchParam() {
     var where = {};
     var or = [];
 
-    _.each(schema.fields, function (field) {
+    _.each(fields, function (field) {
       // NOTICE: Ignore Smart field.
       if (field.isVirtual) { return; }
 
@@ -60,8 +100,11 @@ function ResourcesGetter(model, opts, params) {
         var primaryKeyType = model.primaryKeys[schema.idField].type;
 
         if (primaryKeyType instanceof DataTypes.INTEGER) {
-          q[field.field] = parseInt(params.search, 10) || 0;
-          or.push(q);
+          var value = parseInt(params.search, 10) || 0;
+          if (value) {
+            q[field.field] = value;
+            or.push(q);
+          }
         } else if (primaryKeyType instanceof DataTypes.STRING) {
           columnName = field.columnName || field.field;
           q = opts.sequelize.where(
@@ -105,8 +148,7 @@ function ResourcesGetter(model, opts, params) {
     });
 
     // NOTICE: Handle search on displayed belongsTo
-    _.each(model.associations, function (association) {
-
+    _.each(associations, function (association) {
       if (!fieldNamesRequested ||
         (fieldNamesRequested.indexOf(association.as) !== -1)) {
         if (['HasOne', 'BelongsTo'].indexOf(association.associationType) > -1) {
@@ -119,14 +161,22 @@ function ResourcesGetter(model, opts, params) {
             if (field.reference || field.integration ||
               field.isSearchable === false) { return; }
 
+            if (hasSearchFields && !_.includes(searchAssociationFields,
+              association.as + '.' + field.field)) {
+              return;
+            }
+
             var q = {};
             var columnName = field.columnName || field.field;
             var column = opts.sequelize.col(association.as + '.' + columnName);
 
             if (field.field === schemaAssociation.idField) {
               if (field.type === 'Number') {
-                q = opts.sequelize.where(column, ' = ',
-                  parseInt(params.search, 10) || 0);
+                var value = parseInt(params.search, 10) || 0;
+                if (value) {
+                  q = opts.sequelize.where(column, ' = ',
+                    parseInt(params.search, 10) || 0);
+                }
               } else if (params.search.match(REGEX_UUID)) {
                 q = opts.sequelize.where(column, ' = ', params.search);
               }
@@ -185,6 +235,10 @@ function ResourcesGetter(model, opts, params) {
   }
 
   function getAndCountRecords() {
+    if (hasSearchFields) {
+      selectSearchFields();
+    }
+
     var countOpts = {
       include: queryBuilder.getIncludes(model, fieldNamesRequested),
       where: getWhere()